+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-29 15:17:06
-// Commit: docs(core): update readme and API contracts for core
+// Updated at: 2026-08-29 17:01:50
+// Commit: test(core): implement widget integration tests for core
 
 class coreComponentState {
-  final String commitRef = '4031';
-  final String lastMessage = 'docs(core): update readme and API contracts for core';
-  final DateTime timestamp = DateTime.parse('2026-08-29T15:17:06+05:00');
+  final String commitRef = '4036';
+  final String lastMessage = 'test(core): implement widget integration tests for core';
+  final DateTime timestamp = DateTime.parse('2026-08-29T17:01:50+05:00');
 }